chapter06-llm-agent/langchain/travel-agent-main/travel-agent-main/requirements.txt
langchain==0.2.3
langchain-anthropic==0.1.15
langchain-chroma==0.1.2
langchain-community==0.2.4
langchain-core==0.2.5
langchain-openai==0.1.13
langchain-text-splitters==0.2.1
langchainhub==0.1.20
langdetect==1.0.9
langgraph==0.1.5
numpy==1.26.4
openai==1.35.7
python-dotenv==1.0.1